Fix saveEmp API: change rate field from Integer to Double
The rate parameter accepts decimal values (e.g. 1.15) but was typed as
Integer, causing a type mismatch 400 error. Changed to Double across
entity, history, controller, and service layers.

+++ src/main/java/com/example/workPay/controller/EmployeeController.java
@@ -43,7 +43,7 @@ public class EmployeeController {
             @RequestParam Integer advanceAmount,
             @RequestParam Integer advanceRemaining,
             @RequestParam String salaryType,
-            @RequestParam Integer rate,
+            @RequestParam Double rate,
             @RequestParam Integer clothDoneInMeter,
             @RequestParam(required = false) Integer branchId) {
         Employee employee = Employee.builder()
